+++ src/Pim/Bundle/TransformBundle/Normalizer/Structured/ProductValueNormalizer.php
@@ -39,19 +39,19 @@ class ProductValueNormalizer implements NormalizerInterface, SerializerAwareInte
     public function normalize($entity, $format = null, array $context = [])
     {
         if ($entity->getData() instanceof Collection) {
-            $value = [];
+            $data = [];
             foreach ($entity->getData() as $item) {
-                $value[] = $this->serializer->normalize($item, $format, $context);
-                sort($value);
+                $data[] = $this->serializer->normalize($item, $format, $context);
+                sort($data);
             }
         } else {
-            $value = $this->serializer->normalize($entity->getData(), $format, $context);
+            $data = $this->serializer->normalize($entity->getData(), $format, $context);
         }
 
         return [
             'locale' => $entity->getLocale(),
             'scope'  => $entity->getScope(),
-            'value'  => $value
+            'data'   => $data
         ];
     }